Fix the infinite loop in searching for .annot file
Rewrote the search logic to split out the directory and look under several prefixes (_build, _build/<backend>) for the file. This is a bit of a hack for Mirage, and eventually that system will write a special file to _build giving an index of all the backends to search under. For now though, this works ok.