Make annot work with absolute filenames.
Annot is very subtle when trying to open the given file, for some reason.
When passed an absolute name that's outside of current working directory,
though, it usually fails, with an irritating error message saying that it
can't find this file although the filename is correct.
This happen all the time in VIM when opening a source file from another
repository.
Couldn't stand this anymore so replaced "output an error message and exit"
with "just use the given filename".

Closes #7