Fix for OCaml v4.07.0
At some point between 4.06 and 4.07 the OCaml compiler stopped to
parse linenum indications that do not start at column 0.
The nofake untangle tool fails to meet this requirement, therefore
making annot fails to build with 4.07.
As a workaround, I disabled linenum indication generation.

A better fix would be to fix nofake or better yet, to replace it with
portia which I cured yesterday of the very same condition. :)

Alternatively, you could include the generated source files in the git
repository.